Test `Roughly` properly
Found a couple awkward bugs while testing (mostly around needing to
expand the error margin), but now it's property tested and everything.
